Added posibility to use categorical data (Added category loading to netcdf and csv parsers), which can also be displayed in the parallel coordinates workbench. Further scale and offsets for netcdf should now be used correctly, however this was not yet tested as i have no data with scaled and offset data at hand.

+++ PCViewer/pipelines/scatterplot_renderer.cpp
@@ -227,8 +227,8 @@ void scatterplot_renderer::render(const render_info& info){
         auto render_info = util::vk::initializers::renderPassBeginInfo(framebuffer.render_pass, framebuffer.framebuffer, {0, 0, fb_key.width, fb_key.width}, clear_value);
         vkCmdBeginRenderPass(_render_commands.back(), &render_info, {});
         VkViewport viewport{};
-        viewport.width = fb_key.width;
-        viewport.height = fb_key.width;
+        viewport.width = static_cast<float>(fb_key.width);
+        viewport.height = static_cast<float>(fb_key.width);
         viewport.maxDepth = 1;
         vkCmdSetViewport(_render_commands.back(), 0, 1, &viewport);
         VkRect2D scissor{};
